changes to datatypes on models

+++ property_parks/models.py
@@ -14,8 +14,8 @@ class PropertyPark(models.Model):
   # park_postcode = models.CharField(default=None, max_length=10, null=True, blank=True)
   # postcode_area = models.CharField(default=None, max_length=6, null=True, blank=True)
   # straight_dist_km = models.FloatField(default=None, null=True, blank=True)
-  adjusted_dist_km = models.FloatField(default=None, null=True, blank=True)
-  walking_time_mins = models.FloatField(default=None, null=True, blank=True)
+  adjusted_dist_km = models.DecimalField(default=None, null=True, blank=True, max_digits=3, decimal_places=1)
+  walking_time_mins = models.DecimalField(default=None, null=True, blank=True, max_digits=3, decimal_places=1)
   # driving_time_mins = models.FloatField(default=None, null=True, blank=True)
   # cycling_time_mins = models.FloatField(default=None, null=True, blank=True)
   property_ref = models.ForeignKey(
